+++ regions/_utils/examples.py
